# api/openapi.yaml
openapi: 3.0.3
info:
  title: Launlog API
  version: 0.1.0
  description: |
    Backend API contract for the Launlog laundry POS.

    Date-time values use RFC 3339 with an explicit UTC offset. Business-facing
    timestamps use the Asia/Jakarta timezone (`+07:00`). Date-only values use
    ISO 8601 `YYYY-MM-DD`.

    Rupiah amounts are JSON integers in whole rupiah. Decimal quantities are
    encoded as strings with up to three fractional digits to avoid floating-point
    precision loss.
servers:
  - url: /
    description: Current API server
...
paths:
  /platform/auth/login:
    post:
      operationId: platformLogin
      summary: Authenticate the platform administrator
      tags: [Platform authentication]
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/LoginRequest" } } }
      responses:
        "200": { description: Platform session created., content: { application/json: { schema: { $ref: "#/components/schemas/PlatformAuthSessionResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "413": { $ref: "#/components/responses/RequestTooLarge" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /platform/auth/refresh:
    post:
      operationId: platformRefresh
      summary: Rotate a platform refresh token
      description: Reuse of a consumed token revokes its platform session family.
      tags: [Platform authentication]
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/RefreshRequest" } } }
      responses:
        "200": { description: Platform session rotated., content: { application/json: { schema: { $ref: "#/components/schemas/PlatformAuthSessionResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "413": { $ref: "#/components/responses/RequestTooLarge" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /platform/auth/logout:
    post:
      operationId: platformLogout
      summary: Revoke a platform session family
      tags: [Platform authentication]
      security: [{ PlatformBearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/LogoutRequest" } } }
      responses:
        "204": { description: Platform session revoked. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "413": { $ref: "#/components/responses/RequestTooLarge" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /platform/auth/me:
    get:
      operationId: platformMe
      summary: Get the active platform identity
      tags: [Platform authentication]
      security: [{ PlatformBearerAuth: [] }]
      responses:
        "200": { description: Active platform identity., content: { application/json: { schema: { $ref: "#/components/schemas/PlatformAdmin" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /expense-categories:
    get:
      operationId: listExpenseCategories
      summary: List expense categories in the authenticated business
      tags: [Expenses]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: q
          in: query
          schema: { type: string }
        - name: includeInactive
          in: query
          description: Defaults to false.
          schema: { type: boolean, default: false }
      responses:
        "200": { description: Expense category page., content: { application/json: { schema: { $ref: "#/components/schemas/ExpenseCategoryList" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    post:
      operationId: createExpenseCategory
      summary: Create an expense category
      tags: [Expenses]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ExpenseCategoryInput" } } }
      responses:
        "201": { description: Category created., content: { application/json: { schema: { $ref: "#/components/schemas/ExpenseCategory" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /expense-categories/{categoryId}:
    parameters:
      - name: categoryId
        in: path
        required: true
        schema: { $ref: "#/components/schemas/EntityId" }
    get:
      operationId: getExpenseCategory
      summary: Get an expense category
      tags: [Expenses]
      security: [{ BearerAuth: [] }]
      responses:
        "200": { description: Category details., content: { application/json: { schema: { $ref: "#/components/schemas/ExpenseCategory" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    put:
      operationId: updateExpenseCategory
      summary: Update or deactivate an expense category
      tags: [Expenses]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ExpenseCategoryUpdate" } } }
      responses:
        "200": { description: Category updated., content: { application/json: { schema: { $ref: "#/components/schemas/ExpenseCategory" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /expenses:
    get:
      operationId: listExpenses
      summary: List expenses, defaulting to today in Asia/Jakarta
      description: Filters and reports use expense_at, never created_at. fromDate and toDate are inclusive local dates; omitted dates default to today in Asia/Jakarta. Staff see only currently assigned active outlets.
      tags: [Expenses]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: outletId
          in: query
          schema: { $ref: "#/components/schemas/EntityId" }
        - name: fromDate
          in: query
          schema: { type: string, format: date }
        - name: toDate
          in: query
          schema: { type: string, format: date }
      responses:
        "200": { description: Expense page and applied local date range., content: { application/json: { schema: { $ref: "#/components/schemas/ExpenseList" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    post:
      operationId: createExpense
      summary: Record an expense
      description: Expense amount is positive whole rupiah. Outlet, active category, actor, and tenant are validated transactionally.
      tags: [Expenses]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ExpenseCreateInput" } } }
      responses:
        "201":
          description: Expense recorded with redacted audit data in the same transaction.
          headers: { ETag: { schema: { type: string }, description: Current optimistic-concurrency version. } }
          content: { application/json: { schema: { $ref: "#/components/schemas/Expense" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /expenses/{expenseId}:
    parameters:
      - name: expenseId
        in: path
        required: true
        schema: { $ref: "#/components/schemas/EntityId" }
    get:
      operationId: getExpense
      summary: Get an expense in the authenticated business and outlet scope
      tags: [Expenses]
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Expense details.
          headers: { ETag: { schema: { type: string }, description: Current optimistic-concurrency version. } }
          content: { application/json: { schema: { $ref: "#/components/schemas/Expense" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    put:
      operationId: updateExpense
      summary: Update an expense using optimistic concurrency
      description: If-Match must contain the quoted version returned in the ETag response header. A stale version returns 409 and does not overwrite the intervening update.
      tags: [Expenses]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: If-Match
          in: header
          required: true
          schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ExpenseUpdateInput" } } }
      responses:
        "200":
          description: Expense updated and audited transactionally.
          headers: { ETag: { schema: { type: string } } }
          content: { application/json: { schema: { $ref: "#/components/schemas/Expense" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "428":
          description: If-Match is missing or malformed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /orders:
    get:
      operationId: listOrders
      summary: List orders available to the authenticated user
      tags: [Orders]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: outletId
          in: query
          schema: { $ref: "#/components/schemas/EntityId" }
          description: Optional outlet filter; staff can query only currently assigned active outlets.
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/OrderStatus" }
      responses:
        "200":
          description: Order page.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    post:
      operationId: createOrder
      summary: Create an order and allocate an invoice atomically
      description: Prices are read from active services. Quantity uses NUMERIC(12,3); each line total is exact quantity times snapshot unit price rounded to a whole rupiah using PostgreSQL numeric round (half away from zero), and the order total is the sum of rounded line totals. One optional perfume applies to the whole order. An omitted perfume has NULL ID and name snapshots. Payment status begins UNPAID.
      tags: [Orders]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: Client key scoped to the authenticated business and outlet. Reusing it with the same decoded payload returns the existing order; a different payload returns 409.
          schema: { type: string, minLength: 1, maxLength: 128, pattern: "^[A-Za-z0-9._:-]+$" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrderInput" }
      responses:
        "201":
          description: Order created (or existing order returned for an idempotent replay).
          headers:
            Idempotency-Replayed:
              schema: { type: string, enum: ["true"] }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /orders/{orderId}:
    parameters:
      - name: orderId
        in: path
        required: true
        schema: { $ref: "#/components/schemas/EntityId" }
    get:
      operationId: getOrder
      summary: Get an order and its immutable item snapshots
      tags: [Orders]
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Order details.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /orders/{orderId}/receipt:
    get:
      operationId: getOrderReceipt
      summary: Build an authenticated receipt from historical order snapshots
      description: Item and customer identity values come from receipt-time snapshots. Status, confirmed payment balance, outlet profile, and template are current data. WhatsApp fields provide share content only; no message is sent by the backend.
      tags: [Receipts and settings]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { $ref: "#/components/schemas/EntityId" }
      responses:
        "200":
          description: Receipt and optional WhatsApp share data.
          content: { application/json: { schema: { $ref: "#/components/schemas/ReceiptPayload" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /outlets/{outletId}/receipts/qr/{qrId}:
    get:
      operationId: getReceiptByQR
      summary: Resolve an opaque QR locator to an authorized receipt
      description: QR identifiers are locators, not credentials. Normal bearer authentication, ORDERS_READ permission, tenant ownership, and current outlet assignment are required.
      tags: [Receipts and settings]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/OutletId"
        - name: qrId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Authorized receipt.
          content: { application/json: { schema: { $ref: "#/components/schemas/ReceiptPayload" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /orders/{orderId}/status:
    parameters:
      - name: orderId
        in: path
        required: true
        schema: { $ref: "#/components/schemas/EntityId" }
    patch:
      operationId: transitionOrderStatus
      summary: Advance an order through its allowed lifecycle
      description: "Allowed transitions only: RECEIVED→PROCESSING, PROCESSING→READY_FOR_PICKUP, READY_FOR_PICKUP→COMPLETED. Cancellation uses the separate cancellation operation. COMPLETED and CANCELLED are terminal."
      tags: [Orders]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderStatusTransitionInput" }
      responses:
        "200":
          description: Transition committed with history and audit records.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderLifecycleResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /auth/me:
    get:
      operationId: getCurrentUser
      summary: Get the authenticated user
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user and authorization context.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    PlatformBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Platform JWT access token with the launlog-platform audience."
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token supplied as `Authorization: Bearer <token>`."
...
  schemas:
...
    PlatformAuthSessionResponse:
      type: object
      additionalProperties: false
      required: [tokens, admin]
      properties:
        tokens: { $ref: "#/components/schemas/TokenPair" }
        admin: { $ref: "#/components/schemas/PlatformAdmin" }
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          example: admin@launlog.id
        password:
          type: string
          format: password
          writeOnly: true
          description: Existing credentials are checked as stored. New or changed passwords require 8 to 128 Unicode characters and at most 72 UTF-8 bytes for bcrypt. The character policy is not imposed retroactively at login.
...